[LCR-358] test: 성능 테스트 [perf]
