+++ .lighthouseci/flags-f681b9ec-66a0-4ae3-a666-4dcb4d487d74.json
@@ -1,0 +1,1 @@
+{"chromeFlags":" --headless=new --headless=new"}
